[REM] *: odoo.tools.json.scriptsafe
This utility is clearly widely misunderstood since it's called from a
bunch of Python source files which have nothing to do with `<script>`
tags, some of which only call `loads` for which scriptsafe does
nothing whatsoever.

So remove it, qweb has grown a json layer over it and we can just
merge the two together into qweb's json thingie. The rest can either
use the standard module, or can use the new `odoo.tools.json.simple`
value which avoids exposing any of the complexity in json, it just
exposes loads and dumps.

Just using `json` as the baseline and only injecting a script-safe
serializer inside `<script>` tags has been attempted, however it was
thwarted by the livechat: livechat generates javascript using qweb
templates (not inside a `<script>` tag, just straight javascript) for
the external loader, that context doesn't get recognized as a json
swap location so gets html-escaped JSON in the middle of JS code, but
the livechat loader is *also* injected in `<script>` tags directly so
it can't use its own raw json serialization module.

Making the json serialization module conditional would be an option
but at that point it's not really worth the hassle / complexity, so
keep the current less-efficient-than-it-could-be version.
