+++ addons/base_automation/models/ir_actions_server.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
-from odoo.tools.json import scriptsafe as json_scriptsafe
+from odoo.tools.json import simple as json_simple
 
 from odoo import api, fields, models, _
 from odoo.fields import Domain
@@ -54,7 +54,7 @@ class IrActionsServer(models.Model):
     def _get_eval_context(self, action=None):
         eval_context = super()._get_eval_context(action)
         if action and action.state == "code":
-            eval_context['json'] = json_scriptsafe
+            eval_context['json'] = json_simple
             payload = get_webhook_request_payload()
             if payload:
                 eval_context["payload"] = payload
