+++ odoo/tools/safe_eval/runtime.py
@@ -793,8 +793,7 @@ def _initialize_safe_whitelist():
     safe_whitelist.add_class('odoo.tools.binary.*')
     safe_whitelist.add_class('odoo.tools.convert.xml_import')
     safe_whitelist.add_class('odoo.tools.func.lazy')
-    safe_whitelist.add_class('odoo.tools.json._ScriptSafe')
-    safe_whitelist.add_class('odoo.tools.json.JSON')
+    safe_whitelist.add_class('odoo.tools.json.Simple')
     safe_whitelist.add_class('odoo.tools.profiler.QwebTracker')
     safe_whitelist.add_instance('odoo.tools.misc.ReversedIterable')
     safe_whitelist.add_instance('odoo.tools.query.Query')
